+++ app/javascript/application.js
@@ -1,3 +1,48 @@
 // Configure your import map in config/importmap.rb. Read more: https://github.com/rails/importmap-rails
 import "@hotwired/turbo-rails"
 import "controllers"
+
+
+document.addEventListener('turbo:load', function() {
+  var startBtn = document.getElementById('start-btn');
+  var stopBtn = document.getElementById('stop-btn');
+  var studyRecordId;
+
+  startBtn.addEventListener('click', function() {
+    fetch('/study_records/start', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        'X-CSRF-Token': document.querySelector('[name="csrf-token"]').content
+      }
+    })
+    .then(response => response.json())
+    .then(data => {
+      if (data.success) {
+        studyRecordId = data.study_record_id;
+        startBtn.style.display = 'none';
+        stopBtn.style.display = 'inline';
+      }
+    });
+  });
+
+  stopBtn.addEventListener('click', function() {
+    fetch(`/study_records/${studyRecordId}/stop`, {
+      method: 'PATCH',
+      headers: {
+        'Content-Type': 'application/json',
+        'X-CSRF-Token': document.querySelector('[name="csrf-token"]').content
+      }
+    })
+    .then(response => response.json())
+    .then(data => {
+      if (data.success) {
+        stopBtn.style.display = 'none';
+        startBtn.style.display = 'inline';
+        alert('おつかれさま！！'); 
+      } else {
+        alert('時間測定に失敗しました。');
+      }
+    });
+  });
+});
